[TASK] Backend Web > View: set common preview frame widths
The TYPO3 core features a web preview module with
a simple responsive breakpoint selector.

The default items are cleared and some common
items are added in the following order:

* Computer
* iPhone 5/s
* iPhone 6
* iPhone 6 Plus

These settings could be also set to project specific
breakpoints. This could be something useful and easy
to understand for backend editors.

More details:
https://docs.typo3.org/typo3cms/TSconfigReference/PageTsconfig/Mod/Index.html#web-view-mod-web-view
https://github.com/TYPO3/TYPO3.CMS/blob/5ddb5be78a4fa4dce157c5d8b3647c1d89d3db4c/typo3/sysext/core/Configuration/DefaultConfiguration.php#L1045

Resolves: #24