[FEATURE] CE Youtube: add second "mode" to support custom HTML markup
If you don't use the ViewHelper as self closing HTML tag, the necessary
processed informations (like the complete URL to the YouTube Embed URL)
are handed over to Fluid as variable.
With that it's now possible to write completely custom frontend HTML markup.

Furthermore, the ViewHelper is mentioned in the documentation.

Related: #288
Fixes: #316