[!!!][COMPOSER][TASK] Use neronmoon/scriptsdev to run composer scripts only in composer --dev mode
By using "neronmoon/scriptsdev" it's possible to define composer scripts
only for the non-productive context. All scripts you want to run in a
development context, you have to place under extra.scripts-dev.

With this patch, the folder-structure script is only executed in the development
context. Only the PackageStates.php is generated when executing
`composer install --no-dev` from now on.

This is also the preperation to add development specific scripts
to simplifiy setting up a local TYPO3 distribution