[!!!][COMPOSER][TASK] Update app/composer.json to make TYPO3 9 LTS installable
Since there are still some used 3rd party extensions which
are not compatible to TYPO3 >= 9.0, we have to disable
them per default. Some of them will be added per default,
if they are updated to work with TYPO3 9 LTS.

# composer.json require removals

- apache-solr-for-typo3/solr (install only if you need it)
- causal/image_autoresize
- cmsexperts/unlocalizedcrop due the lack of v9 support
- dmitryd/typo3-dd-googlesitemap is superfluous
- dmitryd/typo3-realurl is superfluous too
- fluidtypo3/vhs must be installed on a per project step
  as not every TYPO3 project needs it as a base
- georgringer/backend-debug is superfluous
- georgringer/deprecationloganalyzer isn't v9 ready
  yet
- georgringer/logging isn't v9 ready yet
- instituteweb/unroll is superfluous
- lochmueller/staticfilecache need
- spooner-web/be_secure_pw isn't v9 ready yet
- t3monitor/t3monitoring_client isn't v9 ready yet

# composer.json require typo3 core packages

The required packages were updated based on the
available one after the TYPO3 upgrade. Features like
felogin must be installed on a per project step.

The amount of per default required core packages
is a purely subjective decision based on my usual
project defaults.

# composer.json suggest

The suggested section was removed. It inflates the file unnecessarily. An alternative will be added within the
documentation of the project.

# composer.json repositories: remove ckeditor-dev

The repo and the package was removed, as this repo
is mainly only relevant if you write CKEditor plugins. If
you need the CKEditor core as a reference for writing
plugins you have to import it in your project by your
own.

# composer.json require-dev removals

- mask/mask was removed due their lack of v9 support
- ichhabrecht/mask-export was removed
- macopedia/mess_detector was removed
- maxserv/yaml_configuration was removed and needs
  some love to work again with v9
- sgalinski/lfeditor was removed and needs some
  love to work again with v9

Related: #340